+++ src/Wolfgang.Etl.FixedWidth/FixedWidthExtractor.cs
@@ -665,11 +665,14 @@ public class FixedWidthExtractor<TRecord> : ExtractorBase<TRecord, FixedWidthRep
     {
         return new FixedWidthReport
         (
-            CurrentItemCount,
-            CurrentSkippedItemCount,
-            CurrentRejectedItemCount,
-            CurrentFilteredLineCount,
-            Interlocked.Read(ref _currentLineNumber)
+            new FixedWidthReportOptions
+            {
+                CurrentCount = CurrentItemCount,
+                CurrentSkippedItemCount = CurrentSkippedItemCount,
+                CurrentRejectedItemCount = CurrentRejectedItemCount,
+                CurrentFilteredLineCount = CurrentFilteredLineCount,
+                CurrentLineNumber = Interlocked.Read(ref _currentLineNumber)
+            }
         );
     }